% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDMICA-package.R
\docType{package}
\name{MDMICA-package}
\alias{MDMICA-package}
\alias{MDMICA}
\title{Independent Component Analysis via Mutual Dependence Measures}
\description{
MDMICA: A package for independent component analysis via mutual dependence measures
}
\details{
The MDMICA package provides independent component analysis methods 
based on mutual dependence measures.
}
\author{
Ze Jin \email{zj58@cornell.edu}
}
